+++ CHANGES.md
@@ -3,6 +3,10 @@
 
 ## 2021
 
+3.8.2 bugfix and code clean for text gravity
+
+3.8.1 bugfix crop and text position calculation
+
 3.8.0 text tool: gravity on/off, point out position
 
 3.7.10 small improvements